requirements-py311_updated.txt
aria2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tensorflow==2.12.0
joblib>=1.1.0
gradio==3.34.0
numba==0.57.0
numpy==1.23.5
scipy
librosa==0.10.0
llvmlite==0.40.0
fairseq @ git+https://github.com/One-sixth/fairseq.git
faiss-cpu==1.7.3
Cython
pydub>=0.25.1
soundfile>=0.12.1
ffmpeg-python>=0.2.0
tensorboardX== 2.6.2.2
Jinja2>=3.1.2
json5
Markdown
matplotlib==3.7.0 
matplotlib-inline>=0.1.7
praat-parselmouth>=0.4.2
Pillow>=9.1.1
resampy>=0.4.2
scikit-learn
tensorboard==2.12.3
tqdm>=4.63.1
tornado>=6.1
Werkzeug>=2.2.3
uc-micro-py>=1.0.1
sympy>=1.11.1
tabulate>=0.8.10
PyYAML>=6.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
fsspec>=2022.11.0
absl-py>=1.2.0
audioread
uvicorn>=0.21.1
colorama>=0.4.5
pyworld==0.3.2
httpx
onnxruntime; sys_platform == 'darwin'
onnxruntime-gpu; sys_platform != 'darwin'
torchcrepe==0.0.22
fastapi==0.88
torchfcpe
ffmpy==0.3.1
python-dotenv>=1.0.0
av


